Base.:* — Method

*(outer::CellPlacement, inner::CellPlacement)

Compose two cell placements. If outer places cell B inside cell A, and inner places cell C inside cell B, then outer * inner describes the placement of cell C inside cell A.

OasisTools.add_cell! — Method

add_cell!(oas, cell)

Add a cell to an OASIS object.

Arguments

oas::Oasis: Your input OASIS object.

cell: Can be a Cell, LazyCell, or simply the name of the new cell you wish to add.

Example

julia> using OasisTools;

julia> filename = joinpath(OasisTools.TESTDATA_DIRECTORY, "boxes.oas");

julia> oas = oasisread(filename)
OASIS file with the following cell hierarchy:
TOP└─ BOTTOM

julia> add_cell!(oas, :NEW)
OASIS file with the following cell hierarchy:
TOP└─ BOTTOMNEW

OasisTools.add_layer! — Method

add_layer!(oas, layer)

Add new layer to your OASIS file.

Arguments

oas::Oasis: Your OASIS file.

layer::Layer: The layer you want to add.

Example

julia> using OasisTools;

julia> filename = joinpath(OasisTools.TESTDATA_DIRECTORY, "nested.oas");

julia> oas = oasisread(filename; lazy = true);

julia> layers(oas)
1-element Vector{Layer}:
 M0 (1/0)

julia> add_layer!(oas, Layer(:V0, 2, 0));

julia> layers(oas)
2-element Vector{Layer}:
 M0 (1/0)
 V0 (2/0)

OasisTools.cell_hierarchy — Method

cell_hierarchy(oas)

Find the cell hierarchy of your OASIS object. Note that lazy-loaded cells will be listed as having no cell placements.

OasisTools.cell_names — Method

cell_names(oas)

Returns a list of all cell names in your OASIS file.

Example

julia> using OasisTools;

julia> filename = joinpath(OasisTools.TESTDATA_DIRECTORY, "boxes.oas");

julia> oas = oasisread(filename);

julia> cell_names(oas)
2-element Vector{Symbol}:
 :BOTTOM
 :TOP

See also cells, name.

OasisTools.cells — Method

cells(oas)

Returns a list of all cells in your OASIS object.

Example

julia> using OasisTools;

julia> filename = joinpath(OasisTools.TESTDATA_DIRECTORY, "polygon.oas");

julia> oas = oasisread(filename);

julia> cells(oas)
1-element Vector{Union{Cell, LazyCell}}:
 Cell TOP with 0 placements and 1 shape

OasisTools.expand — Method

expand(placement::CellPlacement)

Expand a CellPlacement with a repetition into a generator of individual CellPlacement objects, each with repetition = nothing and the repetition offset added to location. If the placement has no repetition, yields a single placement.

OasisTools.layer — Method

layer(oas, shape)
layer(oas, layer_number, datatype_number)

Find the Layer that a given shape or layer/datatype number pair belongs to. Returns nothing if no matching layer is found.

OasisTools.layers — Method

layers(oas)

List all (explicitly named) layers in your OASIS object.

OasisTools.load_all_cells! — Method

load_all_cells!(oas)

Load all LazyCell objects in oas into memory, replacing them with Cell objects. Equivalent to calling load_cell! on every cell.

OasisTools.load_cell! — Method

load_cell!(oas, cell_name)

Load a LazyCell into memory and replace the LazyCell with a corresponding Cell in the input OASIS file.

Example

julia> using OasisTools;

julia> filename = joinpath(OasisTools.TESTDATA_DIRECTORY, "nested.oas");

julia> oas = oasisread(filename; lazy = true);

julia> oas[:TOP]
Lazy cell TOP

julia> load_cell!(oas, :TOP);

julia> oas[:TOP]
Cell TOP with 5 placements and 0 shapes

See also load_cell.

OasisTools.load_cell — Method

load_cell(lazy_cell)

Load a LazyCell into memory.

julia> using OasisTools;

julia> filename = joinpath(OasisTools.TESTDATA_DIRECTORY, "polygon.oas");

julia> oas = oasisread(filename; lazy = true);

julia> lazy_cell = oas[:TOP]
Lazy cell TOP

julia> loaded_cell = load_cell(lazy_cell)
Cell TOP with 0 placements and 1 shape

See also load_cell!.

OasisTools.merge_oases! — Method

merge_oases!(oas, others...)

Update OASIS file with content from the other OASIS files. Using an opinionated plural for 'OASIS'.

OasisTools.merge_oases — Method

merge_oases(oas...)

Merge one or more OASIS files. Using an opinionated plural for 'OASIS'.

Example

julia> using OasisTools;

julia> filename1 = joinpath(OasisTools.TESTDATA_DIRECTORY, "nested.oas");

julia> filename2 = joinpath(OasisTools.TESTDATA_DIRECTORY, "trapezoids.oas");

julia> oas1 = oasisread(filename1; lazy = true)
OASIS file with the following cell hierarchy:
TOP
└─ ?

julia> oas2 = oasisread(filename2)
OASIS file with the following cell hierarchy:
noname

julia> oas = merge_oases(oas1, oas2)
OASIS file with the following cell hierarchy:
TOP
└─ ?noname

OasisTools.name — Method

name(cell)

Name of a cell.

OasisTools.name — Method

name(layer)

Name of a layer.

OasisTools.oasisread — Method

oasisread(filename; kw...)

Read your OASIS file in Julia.

Keyword Arguments

lazy::Bool = false: If set to true, oasisread will use an experimental lazy loader which does not load the contents of the cells into memory until a cell explicitly gets prompted.

Example

julia> using OasisTools;

julia> filename = joinpath(OasisTools.TESTDATA_DIRECTORY, "polygon.oas");

julia> oas = oasisread(filename)
OASIS file with the following cell hierarchy:
TOP

Caveats

There will be bugs.

Properties are not stored.

Backwards-compatible extensions are not supported. You will get an error if your file contains any.

Curvilinear features are not yet supported.

OasisTools.oasiswrite — Method

oasiswrite(filename, oas; bufsize)

Write an Oasis object to an OASIS file.

Arguments

filename::AbstractString: Path of the output file.

oas::Oasis: The OASIS object to write.

Keyword Arguments

bufsize::Integer = 16 * 1024 * 1024: Size of the output buffer in bytes.

OasisTools.placements — Method

placements(cell)

List the placements contained in cell. Not yet supported for LazyCells.

OasisTools.roots — Method

roots(oas)

List cells that are known or presumed to be root cells in the OASIS file.

OasisTools.shapes — Method

shapes(cell)

List the shapes contained in cell. Not yet supported for LazyCells.

OasisTools.show_cells — Method

show_cells(oas; kw...)

Obtain an overview of the cells in your OASIS objects.

Arguments

oas::Oasis: Your OASIS object, loaded with oasisread.

Keyword Arguments

maxdepth = 100: Specify until what maxdepth you'd like to the cell hierarchy to be displayed.

flat = false: If set to true, rather than displaying a hierarchy, show_cells simply lists the names of all cells that can be found in oas. If set to true, the keyword argument maxdepth is ignored.

OasisTools.unit — Method

unit(cell)
unit(oas)

Unit length of a cell or OASIS object, in grid steps per micron.

OasisTools.update_roots! — Method

update_roots!(oas)

Go through the cell hierarchy to figure out what cells are likely to be root cells. May be inaccurate if your OASIS file contains any LazyCell objects, as their placements are unknown.

OasisTools.Cell — Type

struct Cell

Properties

name::Symbol: Name of the cell.

shapes::Vector{Shape}: Lists the shapes, such as polygons and lines, that are contained in the cell.

placements::Vector{CellPlacement}: Lists all other cells that are placed in this cell.

unit::Float64: Unit length.

_root::Bool: Indicates whether the cell is a root cell (i.e., isn't contained in any other cell).

See also LazyCell.

OasisTools.CellHierarchy — Type

struct CellHierarchy

Encodes the cell hierarchy of an OASIS file.

Properties

hierarchy::Dict{Symbol, Vector{Symbol}}. The keys are the cell names, and their values are the (unique) names of the cells that have been placed within them.

OasisTools.CellPlacement — Type

struct CellPlacement

Object encoding the placement of a cell in another cell.

Properties

cellName::Symbol: Name of cell that's being placed.

location::Point{2, Int64}: Where the cell will be placed.

rotation::Float64: Counterclockwise rotation (in degrees) of the cell.

magnification::Float64: Magnification of the cell.

flipped::Bool: Indicates whether or not the cell is reflected (or flipped) around the x-axis. Note: If a cell is flipped and has nonzero rotation, then the flip is applied first, and the rotation is applied second.

repetition: Specifies whether the placement is repeated. If not, repetition = nothing.

OasisTools.Layer — Type

struct Layer

A named layer, associating a human-readable name with a layer number and datatype number (or ranges thereof).

Properties

name::Symbol: Name of the layer.

layerNumber::Interval: Layer number or range of layer numbers.

datatypeNumber::Interval: Datatype number or range of datatype numbers.

OasisTools.LazyCell — Type

struct LazyCell

Lazy-loaded version of a Cell.

Properties

name::Symbol: Name of the cell.

bytes::Vector{UInt8}: Bytes of the corresponding CELL record.

cellnameReferences::Dict{UInt64, Symbol}: To ensure all bytes of the corresponding CELL record are interpretable, a LazyCell stores a list of internal cell name references.

textstringReferences::Dict{UInt64, Symbol}: Internal text string references, stored for the same reason as cellnameReferences.

unit::Float64: Unit length.

_root::Bool: Indicates whether the cell is a root cell (i.e., isn't contained in any other cell).

See also Cell, load_cell!.

OasisTools.Oasis — Type

struct Oasis

Object containing all the data of your OASIS file.

Properties

cells::Vector{Union{LazyCell, Cell}}: Cells in your OASIS file. The cells can either be Cell objects or lazy-loaded LazyCell objects.

layers::Vector{Layer}: Lists all named layers in your OASIS file.

OasisTools.Path — Type

Path(points, width)

A polyline with finite width, or equivalently, a GeometryBasis.LineString with specified width.

OasisTools.PointGridRange — Type

struct PointGridRange(start, nstepx, nstepy, stepx, stepy)

A two-dimensional version of an ordinary range.

Example

PointGridRange((0, 0), 4, 3, (5, 1), (2, -2)) would kind of look like:

                 o
            o      
       o           o
  o           o
         o           o
    o           o
           o
      o

OasisTools.Shape — Type

struct Shape(shape, layerNumber, datatypeNumber, repetition)

Geometric shape (such as a polygon or rectangle) or text.

Properties

shape: The actual shape. If the shape is geometric, then shape::GeometryBasics.GeometryPrimitive{2, Int64}, unless the shape is a path, in which case shape::OasisTools.Path because GeometryBasics doesn't have an appropriate object to encode paths. If the shape is text, then shape::OasisTools.Text.

layerNumber::UInt64: The layer that your shape lives in. You can find the name of the layer using layer.

datatypeNumber::UInt64: The 'datatype' that your shape lives in. To clarify, if your shape lives in (1/0), then datatypeNumber = 0.

repetition: Specifies whether the shape is repeated. If not, repetition = nothing.

OasisTools.plot_cell — Function

plot_cell(cell)

Plot a cell using Makie.

OasisTools.plot_shape! — Function

plot_shape!(ax, shape)

Plot a single Shape onto a Makie Axis. Supported for shapes backed by AbstractGeometry{2, Int64} (rectangles, polygons, etc.).